requirements36.txt
absl-py==0.14.1
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
catboost==1.0.0
cffi==1.15.0
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
FLAML==0.6.0
gast==0.5.2
google-pasta==0.2.0
graphviz==0.17
grpcio==1.41.0
h5py==2.10.0
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazr.uri==1.0.6
lief==0.9.0
lightgbm==3.3.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
redis==3.5.3
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
tenacity==8.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.6
vowpalwabbit==8.11.0
wadllib==1.3.2
Wand==0.5.7
wcwidth==0.1.7
webencodings==0.5
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xcffib==0.5.1
xgboost==1.4.2
xmltodict==0.12.0
zipp==3.5.0
